Nvim: tidies up files for gitsigns
